+++ src/core/Table/Table.tsx
@@ -345,9 +345,7 @@ export const Table = <
       {...getTableProps({
         className: cx(
           'iui-table',
-          {
-            [`iui-${density}`]: density !== 'default',
-          },
+          { [`iui-${density}`]: density !== 'default' },
           className,
         ),
         style,